enable some serious warnings
I found some warnings that seem serious to me, but which are not
enabled by /W3.  Turn them on, and make them errors.  Fix the
code to not fail these warning checks.

Unfortunately, the compiler seems to ignore my attempt to
enable C5205!

+++ GM/VwEdtbrd.cpp
@@ -1675,7 +1675,7 @@ void CBrdEditView::OnUpdateEditPasteBitmapFromFile(CCmdUI* pCmdUI)
 
 //DFM 19991014...
 //GetKeyState()
-static const short KEY_STATE_VALUE = (short)0x8000;
+static const short KEY_STATE_VALUE = static_cast<short>(static_cast<unsigned short>(0x8000));
 
 void CBrdEditView::HandleKeyDown ()
 {
